+++ packages/cli/CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### Patch Changes
 
-- Deploy test
+- Updated dependencies [2a0aaa9]
+  - @openfn/compiler@0.0.35
+  - @openfn/runtime@0.0.29
+  - @openfn/deploy@0.2.3
+  - @openfn/logger@0.0.16
 
 ## 0.2.3