+++ packages/runtime-manager/CHANGELOG.md
@@ -1,5 +1,14 @@
 # runtime-manager
 
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [2a0aaa9]
+  - @openfn/compiler@0.0.35
+  - @openfn/runtime@0.0.29
+  - @openfn/logger@0.0.16
+
 ## 0.0.37
 
 ### Patch Changes